requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.7.0
blobfile==2.0.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.1
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.9.0
Flask==3.0.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
install==1.3.5
ipykernel==6.20.1
ipython==8.10.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.24.1
openai==1.9.0
overrides==7.4.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.12
platformdirs==2.6.2
plotly==5.12.0
prometheus-client==0.19.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.17
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2022.10.31
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
safetensors==0.4.1
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tenacity==8.1.0
terminado==0.18.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tinycss2==1.2.1
tokenizers==0.13.2
tornado==6.3.3
tqdm==4.64.1
traitlets==5.8.1
transformers==4.30.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yarg==0.1.9
yarl==1.8.2
